chore: Moves Modal to the components folder (#14130)
