+++ superset-frontend/spec/javascripts/dashboard/components/PropertiesModal_spec.jsx
@@ -27,7 +27,7 @@ import {
   ThemeProvider,
 } from '@superset-ui/core';
 
-import Modal from 'src/common/components/Modal';
+import Modal from 'src/components/Modal';
 import PropertiesModal from 'src/dashboard/components/PropertiesModal';
 import { mockStore } from 'spec/fixtures/mockStore';
