+++ superset-frontend/src/SqlLab/components/SaveDatasetModal.tsx
@@ -21,7 +21,7 @@ import React, { FunctionComponent } from 'react';
 import { AutoCompleteProps } from 'antd/lib/auto-complete';
 import { Radio } from 'src/common/components/Radio';
 import { AutoComplete, Input } from 'src/common/components';
-import StyledModal from 'src/common/components/Modal';
+import StyledModal from 'src/components/Modal';
 import Button from 'src/components/Button';
 import { styled, t } from '@superset-ui/core';
 import { RadioChangeEvent } from 'antd/lib/radio';
