+++ superset-frontend/src/explore/components/controls/VizTypeControl.jsx
@@ -22,7 +22,7 @@ import { Row, Col, FormControl } from 'react-bootstrap';
 import { Behavior, t, getChartMetadataRegistry } from '@superset-ui/core';
 import { useDynamicPluginContext } from 'src/components/DynamicPlugins';
 import { Tooltip } from 'src/common/components/Tooltip';
-import Modal from 'src/common/components/Modal';
+import Modal from 'src/components/Modal';
 import Label from 'src/components/Label';
 import ControlHeader from '../ControlHeader';
 import './VizTypeControl.less';
