+++ superset-frontend/src/views/CRUD/data/savedquery/SavedQueryPreviewModal.tsx
@@ -18,7 +18,7 @@
  */
 import React, { FunctionComponent } from 'react';
 import { styled, t } from '@superset-ui/core';
-import Modal from 'src/common/components/Modal';
+import Modal from 'src/components/Modal';
 import Button from 'src/components/Button';
 import SyntaxHighlighterCopy from 'src/views/CRUD/data/components/SyntaxHighlighterCopy';
 import withToasts, { ToastProps } from 'src/messageToasts/enhancers/withToasts';
